+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This file is a history of the changes made to @idearium/cli.
 
 ## Unreleased
 
+### Added
+
+-   `c d image prune` command.
+
 ## v5.1.0-beta.3 - 2022-08-09
 
 ### Fixed